requirements.txt
fastapi>=0.63,<0.64.0
jinja2
python-dotenv
python-multipart
requests
uvicorn